+++ C#_Version/DeconEngine/HornTransform/ChiSqFitScorer.cs
@@ -16,7 +16,7 @@ namespace Engine.HornTransform
         /// <param name="minIntensityForScore">minimum intensity for score</param>
         /// <param name="pointsUsed">number of points used</param>
         /// <param name="debug">debug output flag</param>
-        public override double FitScore(PeakData peakData, short chargeState, clsPeak peak, double mzDelta,
+        public override double FitScore(PeakData peakData, int chargeState, clsPeak peak, double mzDelta,
             double minIntensityForScore, out int pointsUsed, bool debug = false)
         {
             pointsUsed = 0;
@@ -65,7 +65,7 @@ namespace Engine.HornTransform
         /// </param>
         /// <param name="minIntensityForScore">minimum intensity for score</param>
         /// <param name="debug">prints debugging information if this is set to true.</param>
-        public override double FitScore(PeakData peakData, short chargeState, double intensityNormalizer, double mzDelta,
+        public override double FitScore(PeakData peakData, int chargeState, double intensityNormalizer, double mzDelta,
             double minIntensityForScore, bool debug = false)
         {
             var numPoints = TheoreticalDistMzs.Count;
