+++ C#_Version/DeconEngine/HornTransform/IsotopeFitRecord.cs
@@ -32,7 +32,7 @@ namespace Engine.HornTransform
         /// <summary>
         ///     charge state
         /// </summary>
-        public short ChargeState;
+        public int ChargeState;
 
         /// <summary>
         ///     difference between obsered m/z and m/z from theoretical distribution of composition from Averagine
