+++ C#_Version/DeconEngine/HornTransform/PeakFitScorer.cs
@@ -17,7 +17,7 @@ namespace Engine.HornTransform
         /// <param name="minIntensityForScore">minimum intensity for score</param>
         /// <param name="pointsUsed">number of points used</param>
         /// <param name="debug">debug output flag</param>
-        public override double FitScore(PeakData peakData, short chargeState, clsPeak peak, double mzDelta,
+        public override double FitScore(PeakData peakData, int chargeState, clsPeak peak, double mzDelta,
             double minIntensityForScore, out int pointsUsed, bool debug = false)
         {
             pointsUsed = 0;
@@ -92,7 +92,7 @@ namespace Engine.HornTransform
         /// </param>
         /// <param name="minIntensityForScore">minimum intensity for score</param>
         /// <param name="debug">prints debugging information if this is set to true.</param>
-        public override double FitScore(PeakData peakData, short chargeState, double intensityNormalizer, double mzDelta,
+        public override double FitScore(PeakData peakData, int chargeState, double intensityNormalizer, double mzDelta,
             double minIntensityForScore, bool debug = false)
         {
             throw new System.Exception("Don't Ever come into this FitScore overload in PeakFit");
